Updated netCDF payload to ignore fields specified in the ignore field in institution mapping, also wrote examples againg, and updated knmi mapping to dispalay lat,lon in correct order. And moved netCDF scripts in to netCDF module.
