+++ examples/20221231_meta.json
@@ -1,6 +1,7 @@
 {
     "properties": {
         "metadata_id": "Royal Netherlands Meteorological Institute (KNMI):fbfad5b9-1dd2-425e-bb35-c96386380c0e",
+        "data_id": "placeholder",
         "creator_url": "http://data.knmi.nl",
         "creator_email": "datacentrum@knmi.nl",
         "keywords": "temperature, pressure, relative humidity, visibility, wind speed, wind direction, wind gust",
@@ -12,7 +13,7 @@
         "history": "File created from KMDS ASCII file. ",
         "source": "Royal Netherlands Meteorological Institute (KNMI)",
         "datetime": "2022-12-31T00:00:00.000000000",
-        "pubtime": "2023-08-29T08:43:38.000000000Z"
+        "pubtime": "2023-09-04T12:51:43.000000000Z"
     },
     "links": [
         {
@@ -23,18 +24,18 @@
     "geometry": {
         "type": "Point",
         "coordinates": [
-            10.0,
-            60.0
+            60.0,
+            10.0
         ]
     },
     "type": "Feature",
     "version": "v4.0",
     "content": {
         "encoding": "utf-8",
-        "standard_name": "latitude",
-        "unit": "degrees_north",
-        "size": 15,
-        "value": "54.325666666667"
+        "standard_name": "wind_from_direction",
+        "unit": "degree",
+        "size": 5,
+        "value": "214.5"
     },
-    "id": "38c0bd1f-4f2c-4595-99b4-2006a85ae318"
+    "id": "3d61797c-700f-4443-b265-0579dc1ffe00"
 }
