+++ examples/air_temperature_gullingen_skisenter-parent_meta.json
@@ -1,6 +1,7 @@
 {
     "properties": {
         "metadata_id": "no.met:8a0744b7-f0dc-4929-8700-f1285e76df36",
+        "data_id": "placeholder",
         "title": "Air temperature observations from weather station Gullingen , ROGALAND, Norge, (WIGOS ID: 0-578-0-46220, station ID: 46220).",
         "institution": "Norwegian Meteorological Institute (MET Norway)",
         "Conventions": "CF-1.10, ACDD-1.3",
@@ -14,7 +15,7 @@
         "license": "http://spdx.org/licenses/CC-BY-4.0(CC-BY-4.0)",
         "access_constraint": "Open",
         "datetime": "2023-06-13T14:00:00.000000000",
-        "pubtime": "2023-08-29T08:43:38.000000000Z"
+        "pubtime": "2023-09-04T12:51:43.000000000Z"
     },
     "links": [
         {
@@ -38,5 +39,5 @@
         "size": 18,
         "value": "20.700000762939453"
     },
-    "id": "3dae3e50-2b88-463d-ad84-ee68ff2c1745"
+    "id": "ff6ddfa5-4020-47c2-a791-3dc6834fd0e6"
 }
