+++ schemas/netcdf_to_e_soh_message_knmi.json
@@ -10,8 +10,8 @@
 					"coordinates": {
 						"inpt_type": "list",
 						"fields": [
-							"max-x",
-							"max-y"
+							"max-y",
+							"max-x"
 						]
 					},
 					"inpt_type": "multi"
@@ -104,5 +104,12 @@
 			]
 		}
 	},
-	"description": "This json defines what fields to take from a netCDF file in to a E-SOH MQTT message. Fields in the translation have different names between the netCDF and MQTT message. persistant_fields are fields that can be directly extrated from the netCDF. This file is defined for netCDF created at METno."
+	"description": "This json defines what fields to take from a netCDF file in to a E-SOH MQTT message. Fields in the translation have different names between the netCDF and MQTT message. persistant_fields are fields that can be directly extrated from the netCDF. This file is defined for netCDF created at METno.",
+	"ignore": [
+		"time",
+		"stationname",
+		"lat",
+		"lon",
+		"height"
+	]
 }
